requirements.txt
aiogram
python-dotenv
requests
apscheduler==3.10.4
alembic
sqlalchemy
aiosqlite
aiohttp==3.8.5
aiocache